Build CodeInput, completes styling for mobile
