+++ src/app/Components/BackButton/BackButton.module.css
@@ -1,17 +1,8 @@
 .button {
-  background: transparent;
   position: relative;
   z-index: 1;
   width: 7.5rem;
   height: 3rem;
   left: 0.5rem;
   top: 0.5rem;
-  border: 2px solid #92bde3;
-  border-radius: 3px;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 1rem;
-  line-height: normal;
-  letter-spacing: 0.1em;
-  color: #92bde3;
 }
