+++ src/app/Components/CodeInput/CodeInput.module.css
@@ -1,0 +1,34 @@
+.code-input-section {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  height: 20rem;
+  text-align: center;
+}
+
+.welcome-headline {
+  font-size: 1.6rem;
+  padding: 0rem 0.25rem;
+}
+
+.input-div {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+}
+
+.code-input {
+  height: 3rem;
+  width: 21.25rem;
+  background: #fffefa;
+  border: 2px solid #3054a1;
+  border-radius: 3px;
+  text-align: center;
+}
+
+.submit-btn {
+  width: 11.5rem;
+  height: 3rem;
+}
