+++ src/app/globals.css
@@ -20,13 +20,25 @@ html {
 }
 
 body {
-  color: #7D7F82;
-  background: #FFFEFA;
+  color: #7d7f82;
+  background: #fffefa;
 }
 
 a {
-  color: inherit;
-  text-decoration: none;
+  text-decoration-line: underline;
+  color: #5497cf;
+}
+
+button {
+  background: #FFFEFA;
+  border: 2px solid #92bde3;
+  border-radius: 3px;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 1rem;
+  line-height: normal;
+  letter-spacing: 0.1em;
+  color: #92bde3;
 }
 
 h1 {
